requirements.txt
boto3>=1.28.0
pytest>=7.0
moto>=4.0.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
pytest-mock>=3.10